pyproject.toml
[project]
name = "trmnl-gcal"
version = "0.1.0"
description = "TRMNL Private Plugin for Google Calendar"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "google-cloud-secret-manager>=2.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
